:sparkles: Update Github stars fetching mechanism so that entries can be sorted by stars (#59)
This PR changes the ordering of entries in the component catalog so that
the entries with the most stars appear first.

The simplest way to do this requires the stars to be already included in
the components' frontmatters. Therefore, the `fetch_stars.py` script now
loads the values into the frontmatters.

This feature also requires some additional features:
- Until now, the frontmatters of the components did not use the
"canonical" markdown syntax. This had to be updated for all existing
components and for the automatic component generation workflow, in
particular for:
  - strings (here, quotation marks are not required)
- arrays (using enumeration points starting with `-` is more standard
than using square brackets)
- A pre-commit hook has been added that removes fetched star counts if
the page has been built locally so that the stars never make it into the
repository.